@@ -1,5 +1,10 @@
 module uk.ac.bris.cs/gameoflife
 
-go 1.17
+go 1.24.0
 
 require github.com/veandco/go-sdl2 v0.4.40
+
+require (
+	github.com/aclements/go-moremath v0.0.0-20210112150236-f10218a38794 // indirect
+	golang.org/x/perf v0.0.0-20251112180420-cfbd823f7301 // indirect
+)
